Use tox + Github Actions for tests
Actually Travis doesn't make build logs public, which is not cool for contributing, and it doesn't make it easy to run tests locally and see results.

With this commit I switch it to tox which makes it easier to run tests locally and see coverage, plus Github Actions.

Also tests are moved outside of the source directory